feat: add market name to the order cancellation screen
